Added user management and admin list
Added some work on the user management, so a signed in administrator can now delete, edit or view a user. This will only be possible if the signed in user is a admin, which will be checked by the new User::is_admin() method. The fore-mentioned method will check if the privilege_level of the $session->user_id to the signed in user is 0 or 1, where 0 is a user and 1 is a admin.  

The list shown on users.php will change whether the user is a user or an admin, and they should have different looks when we come to the front-end, with the admin one being more simple.

As mentioned there were added a new method in the User-class that checks if the user is an admin, which should make it easy to check.

The users that is not logged in can no longer look at the user-list.