Added logging of user-activity
The log_user_activity now logs when something is created, updated or deleted. The change is added to the new table that have been created, named 'user_activity'.

A change is added to the CRUD methods by adding a call to the log_user_activity() and sends in the type of action, the type of object and the id for the object that have been changed/deleted/created. The user-id of the signed in user is added inside of the method if someone is logged in, when registering a user it will show the id of the user that is made.