Added the add_friend function
It is placed in the user-list at the moment, the user-list for users will not be there later, but much of it can get transported to the places they should be later easily.

At the moment it only adds the row to the database and the is_friend only checks if the "Friend request" was sent, need to modify it later when the receiver can accept or deny one. The user who sent the friend request should also be able to cancel it.