Search feature on the game-list
Added a search feature to the game list, at the moment it can search after games with the genre, title and creator at the same time. It uses a method called find_game() where it writes up a SQL command that uses the three parameters.

Before merge i will make a simple search on the user-list too. The is things that can be done in the future that will make it better, maybe Davids recommendation, giving each games multiple genres etc.