Placed games on the gamepage
Games now plays on the gamepage instead of their base Php-file, currently this shits on the games that requires sprites and assets, but those that only hold a php file will work fine, we'll fix this later. Will send game_id over _GET instead of link, but does this tomorrow.