Added ajax to userlist and added js-file
Moved the java script over to functions.js in the js-folder, to clean up the page, placed the script in the body to secure better speed. Added ajax to the userlist and new javascript-function to the userlist called update_userlist. 

Later in this branch ajax will be added to the edit pages, the game-list and the upload page