Added ajax to game list and changed searching
After discussion we it was decided to search by using one input text field, had to change the back-end of it. Now searches is made with every placed in character, or changed one of the list-item, changed the search-functions and cleaned up a little, removed the admin_gamelist and added the delete user function to the game list instead, only visible for users, new function in the function.js. More ajax to add tomorrow.