Added search-for-friend list and security changes
The ability for users to search for people to add as friends now lies in the profile page, it will of course be heavily modified later when it added to the front-end. Did some security changes, removed the block for special characters, real_escape_string helps stopping SQL-injections, changed max characters to 50 from 30, and added call on escape_string some places. Created friend_search stat is used when the feature is used, and placed ajax on the list and button.