Added accept and decline friend request
Fixed delete() by adding $session, as it uses it. Added some messages to better explain some of the pages. Added two buttons where the user can accept and decline friend-requests, but at the moment first object on the list will be updated , even trough it will work correctly in the back-end. Removed unnecessary call to the database in the friend_request_handler() . Changed the friend_request_handler.php to add a message to the user when updated. Added <mark> on the search-words in friend-search and the user-list.