Changed the file-structure and worked more on the chat
Added 4 new folders inside of includes. Classes will keep all the classes. Helpers will hold all the php-files that don't fit in classes or process, like config.php and database.php. Process will contain the php-files that does something, mostly called with ajax. Views will contain the files that shows up on the page, example is lists, content, footer and header. The pages like index.php will be in the main directory.

Created userchat.php to be able place it anywhere, changed the table so that which game the comment is done on, will call on only the comments made on the side of the game.

Had to change a lot of the things file-paths, the init on the files in the new folders had to be changed.