Added the new header, foreign keys and more
Moved head.php:
Created head.php and placed it in helpers it needs to be called on pages that the header should not be placed on.

Foreign Keys on database:
Added all the keys that the database, there is still some cleaning to do on the names, used cascade to ensure no problems when users is removed, need to use cascade more later.

Added Header Front-end:
Davids front-end were added,  had to do quite a lot of work for the back-end to fit it in, removed the movement bar that was there before.

Changed File-Structure:
Placed the assets in the assets-folder and changed the includes for main.css and main.js to fit the changes.

- Cleaned some folders and files
- Added some files that I will use later
- Made some large changes to the database
- Moved the JavaScript from functions.js to main.js
- Got the page name from the URI to decide which of the options should be '-active'