Worked on the friend-chat and PHPdoc on Db_object
The friend-chat is now up and ready to be worked on more, currently used the controller to what room as a radio-button behind the friend, clicking on it will give you the chat based on the friends id at the moment.
This will be changed later to both the players id with a # in the middle to make it unique. 

Started the great documenting by doing the Db_object one , plan to finish the classes tomorrow.