More PHP-Doc and small changes
Worked on the doc in the classes. The page-names is now stored at the beginning of the main-pages, is currently used whenever it have to check what page its on in the first navigation. made the chatId input hidden, the css should be placed in main.css later. Did some small changes in the database.