Added changes to the chat
Room id now gets changed whenever one of the chats is crossed out, goes to zero when there is none left, uses .getAttribute() to collect the "value" of div, chat messages is seperated by days with lines, there is still the problem with the chat shrinking if the  long angry message is taken away, David can fix it quickly on the next meeting.

Left to do:
- Active or not circle thingy.
- Make message-input disperse of there is zero chats.
- The chats need to be able to be activated with a 'message them' button and deactivated when the X is clicked, need another colon on the chat_room table.
- Need to start at the end of the chat.
- Used a innerText to set the huge title--username-thingy.

That's all for now.